+++ src/main/java/org/jabref/logic/importer/fileformat/BibtexImporter.java
@@ -1,9 +1,10 @@
 package org.jabref.logic.importer.fileformat;
 
+import java.io.BufferedInputStream;
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStream;
 import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Objects;
@@ -16,10 +17,15 @@ import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.util.StandardFileType;
 import org.jabref.model.util.FileUpdateMonitor;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 /**
  * This is a full class to read .bib files. It is used for <code>--import</code> and <code>--importToOpen </code>, too.
  */
 public class BibtexImporter extends Importer {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(BibtexImporter.class);
 
     // Signature written at the top of the .bib file in earlier versions.
     private static final String SIGNATURE = "This file was created with JabRef";
@@ -44,27 +50,39 @@ public class BibtexImporter extends Importer {
 
     @Override
     public ParserResult importDatabase(Path filePath) throws IOException {
-        // We want to check if there is a JabRef signature in the file, because that would tell us
-        // which character encoding is used. However, to read the signature we must be using a compatible
-        // encoding in the first place. Since the signature doesn't contain any fancy characters, we can
-        // read it regardless of encoding, with either UTF-8 or UTF-16. That's the hypothesis, at any rate.
-        // 8 bit is most likely, so we try that first:
-        Optional<Charset> suppliedEncoding;
-        try (BufferedReader reader = Files.newBufferedReader(filePath)) {
-            reader.mark(8192);
-            suppliedEncoding = getSuppliedEncoding(reader);
-            if (!suppliedEncoding.isPresent() || suppliedEncoding.equals(StandardCharsets.UTF_8)) {
-                // read with UTF-8 in case no encoding string is present - or the given charset is UTF-8
-                reader.reset();
-                return this.importDatabase(reader);
-            }
-        }
-        try (BufferedReader reader = Files.newBufferedReader(filePath, suppliedEncoding.get())) {
-            return this.importDatabase(reader);
+        // We want to check if there is a JabRef encoding heading in the file, because that would tell us
+        // which character encoding is used.
+
+        // In general, we have to use InputStream and not a Reader, because a Reader requires an encoding specification.
+        // We do not want to do a byte-by-byte reading or doing wild try/catch magic.
+        // We therefore use a charset detection library and then read JabRefs "% Encoding" mark
+
+        Charset detectedCharset;
+        try (InputStream inputStream = Files.newInputStream(filePath)) {
+            BufferedInputStream bufferedInputStream = new BufferedInputStream(inputStream);
+            bufferedInputStream.mark(8192);
+            detectedCharset = getCharset(bufferedInputStream);
+            bufferedInputStream.reset();
         }
 
+        Charset encoding;
+        try (BufferedReader reader = Files.newBufferedReader(filePath, detectedCharset)) {
+            Optional<Charset> suppliedEncoding = getSuppliedEncoding(reader);
+            // in case no encoding information is present, use the detected one
+            encoding = suppliedEncoding.orElse(detectedCharset);
+        }
+
+        try (BufferedReader reader = Files.newBufferedReader(filePath, encoding)) {
+            ParserResult parserResult = this.importDatabase(reader);
+            parserResult.getMetaData().setEncoding(encoding);
+            return parserResult;
+        }
     }
 
+    /**
+     * This method does not set the metadata encoding information. The caller needs to set the encoding of the supplied
+     * reader manually to the meta data
+     */
     @Override
     public ParserResult importDatabase(BufferedReader reader) throws IOException {
         return new BibtexParser(importFormatPreferences, fileMonitor).parse(reader);
@@ -122,7 +140,7 @@ public class BibtexImporter extends Importer {
                 }
             }
         } catch (IOException ignored) {
-            // Ignored
+            LOGGER.error("Supplied encoding could not be determined", ignored);
         }
         return Optional.empty();
     }